move GetV6TrainingData() to selfplay/game.cc and simplify it
