+++ src/mcts/node.h
@@ -36,10 +36,6 @@
 #include "chess/board.h"
 #include "chess/callbacks.h"
 #include "chess/position.h"
-#include "neural/cache.h"
-#include "neural/encoder.h"
-#include "neural/writer.h"
-#include "proto/net.pb.h"
 #include "utils/mutex.h"
 
 namespace lczero {
@@ -242,13 +238,6 @@ class Node {
   // Calculates the full depth if new depth is larger, updates it, returns
   // in depth parameter, and returns true if it was indeed updated.
   bool UpdateFullDepth(uint16_t* depth);
-
-  V6TrainingData GetV6TrainingData(
-      GameResult result, const PositionHistory& history,
-      FillEmptyHistory fill_empty_history,
-      pblczero::NetworkFormat::InputFormat input_format, Eval best_eval,
-      Eval played_eval, bool best_is_proven, Move best_move, Move played_move,
-      const NNCacheLock& nneval, float temp) const;
 
   // Returns range for iterating over edges.
   ConstIterator Edges() const;
