+++ src/selfplay/game.h
@@ -34,6 +34,7 @@
 #include "mcts/stoppers/stoppers.h"
 #include "neural/cache.h"
 #include "neural/network.h"
+#include "neural/writer.h"
 #include "utils/optionsparser.h"
 
 namespace lczero {
@@ -96,6 +97,11 @@ class SelfPlayGame {
   uint64_t nodes_total_ = 0;
 
  private:
+  V6TrainingData GetV6TrainingData(
+      const NodeTree& tree, pblczero::NetworkFormat::InputFormat input_format,
+      Eval best_eval, Eval played_eval, bool best_is_proven, Move best_move,
+      Move played_move) const;
+
   // options_[0] is for white player, [1] for black.
   PlayerOptions options_[2];
   // Node tree for player1 and player2. If the tree is shared between players,
